Letting the user input by STDIN
Usualy UNIX app let the user provide arguments through the STDIN when no arguments are pass to the application.

Ruby allow use to read from STDIN with this const.

Also is a good practice to inform the user about this because the blinking cursor is not enough information about what to do in this sutiations.