Be aware of [non]destructive actions in our app
Be aware of the possible destructive actions our app can do, that way we can provide information and default values to it, making the exeprience for the user more graceful.

In our example our backup will warn the user when there is already a backup file with the same name. To avoid this warning message the user must pass an option.